Revisions and bug fix
Revisions to funtions and implementation. Code had bug that caused results for large datasets to not be accumulated. Only the last batch of data was evaluated and results captured. Also, csv file results were never written for the first dataset to be evaluated because that loop created the files rather than write results. Corrected that by created csv files up front.